Added implementation of first TO DO from my ML notes.
First TO DO I had noted was implemnting a simple cost function
computation and visualization.
Given some data, we generate completely random linear equations,
and try to find which of these generated lines fits best the original,
true line. These randomly generated linear equations are actually
pretending to be model generated by the learning process.
The point of this is to develop a bit of intuition about the inner
workings of cost function, as well as simply write your own code.
A plot example is attached.