Adding gradient descent implementation for linear regression.
This initial implementation of gradient descent for linear regression
does what it was intended to do.
However, there is a number of things to improve.

1. All of vector operations which are currently done in for loops
can be switched out for one-liners.

2. alpha is suspiciously low. I initially tried running it with 0.01,
but it started diverging massively. Then on the other hand, I can see
some other people online are complaining of even smaller alpha values
(10^-5, even lower), so it is probably OK.

3. break clause uses latest cost function to determine whether to
continue running grad desc, or break. If latest cost is below 0.01,
we proclaim the used model as good enough, and exit grad desc.
A couple of issues here :
   -0.01 is arbitrary - what if you have a different model with
vastly different costs
   -you have to exit by taking into account how the cost behaved in
previous iterations

4. General code cleanup