ML:Playground/NN_model_heart_disease/sol.py
Pushing first version of a NN trained on the heart disease dataset.

This initial version uses a 3 layer NN : one ingestion layer made
up of 13 nodes (because we have 13 variables), one hidden layer made
up of 8, and one output layer of 5 nodes.
Input, and hidden, layer use ReLU activation, all output nodes use
softmax activation layer.

Dataset had to cleaned up a bit because it had a couple of NaN
values in it causing issues during loss computation.
After this was cleaned up, model training started working ok.

This initial implementation "hides" the softmax activation function
in the last layer, and uses 5 nodes in the output layer because you
have 5 possible output layers.

To make a better implementation of this you need to remove softmax
from the output layer, and instead use linear activation function.
Once you've trained such a model, when you are trying to predict the
ouputs of a new dataset, you can apply softmax manually to the output
of the last layer. However, even without applying softmax you can
determine which category is the most probable one for the given data.
Softmax is basically just a way of reducing the given values from a wide
range of values, to a [0,1] range.