ML:Playground:NN_heart_disease_model
Tried varying existing data by +/-10% randomly, and the results are a bit
better in sense of cosistency of performance across all data sets
(training, test and CV sets), but the accuracy is horrible - it
generally sits in mid 50% range.
This is most likely because I forgot to do the feature scaling/normalization.
Oops.