No code changes in this commit, just some explanations in the commit
Previous commit pushed the first version where the code successfully assigns
points to closest cluster centroids, as well as adjusts the position of the
cluster centroids based on the newly assigned datapoints.

A lot of stuff needs to be fixed :
-too much list/np.array manipulation - a lot of lists or np.array are being manually
manipulated from lists to np.arrays and vice versa because I find various operations
to be easier using either of these forms. Flatten this to only one of these.

-recheck distances - I didn't really check if my L2_norm computations are ok.
Need to recheck this manually and using some existing L2_norm implementation (probably
exists in numpy).

-the whole algorithm needs to be made iterative - k-means is an iterative process, your
previous commit is basically one iteration of k-means.

-after trimming all of the debug fat, some nice, succint code chunks will probably
be candidates for functions

Primary focus for the next commits is :
-scrutinize L2_norm function
-list/np.array simplification